+++ json_annotation/lib/src/json_serializable.g.dart
@@ -16,24 +16,26 @@ JsonSerializable _$JsonSerializableFromJson(Map<String, dynamic> json) {
       'disallow_unrecognized_keys',
       'explicit_to_json',
       'field_rename',
+      'ignore_unannotated',
       'include_if_null',
       'nullable',
     ]);
     final val = JsonSerializable(
-        anyMap: $checkedConvert(json, 'any_map', (v) => v as bool),
-        checked: $checkedConvert(json, 'checked', (v) => v as bool),
-        createFactory:
-            $checkedConvert(json, 'create_factory', (v) => v as bool),
-        createToJson: $checkedConvert(json, 'create_to_json', (v) => v as bool),
-        disallowUnrecognizedKeys: $checkedConvert(
-            json, 'disallow_unrecognized_keys', (v) => v as bool),
-        explicitToJson:
-            $checkedConvert(json, 'explicit_to_json', (v) => v as bool),
-        fieldRename: $checkedConvert(json, 'field_rename',
-            (v) => _$enumDecodeNullable(_$FieldRenameEnumMap, v)),
-        includeIfNull:
-            $checkedConvert(json, 'include_if_null', (v) => v as bool),
-        nullable: $checkedConvert(json, 'nullable', (v) => v as bool));
+      anyMap: $checkedConvert(json, 'any_map', (v) => v as bool),
+      checked: $checkedConvert(json, 'checked', (v) => v as bool),
+      createFactory: $checkedConvert(json, 'create_factory', (v) => v as bool),
+      createToJson: $checkedConvert(json, 'create_to_json', (v) => v as bool),
+      disallowUnrecognizedKeys:
+          $checkedConvert(json, 'disallow_unrecognized_keys', (v) => v as bool),
+      explicitToJson:
+          $checkedConvert(json, 'explicit_to_json', (v) => v as bool),
+      fieldRename: $checkedConvert(json, 'field_rename',
+          (v) => _$enumDecodeNullable(_$FieldRenameEnumMap, v)),
+      ignoreUnannotated:
+          $checkedConvert(json, 'ignore_unannotated', (v) => v as bool),
+      includeIfNull: $checkedConvert(json, 'include_if_null', (v) => v as bool),
+      nullable: $checkedConvert(json, 'nullable', (v) => v as bool),
+    );
     return val;
   }, fieldKeyMap: const {
     'anyMap': 'any_map',
@@ -42,6 +44,7 @@ JsonSerializable _$JsonSerializableFromJson(Map<String, dynamic> json) {
     'disallowUnrecognizedKeys': 'disallow_unrecognized_keys',
     'explicitToJson': 'explicit_to_json',
     'fieldRename': 'field_rename',
+    'ignoreUnannotated': 'ignore_unannotated',
     'includeIfNull': 'include_if_null',
   });
 }
@@ -55,6 +58,7 @@ Map<String, dynamic> _$JsonSerializableToJson(JsonSerializable instance) =>
       'disallow_unrecognized_keys': instance.disallowUnrecognizedKeys,
       'explicit_to_json': instance.explicitToJson,
       'field_rename': _$FieldRenameEnumMap[instance.fieldRename],
+      'ignore_unannotated': instance.ignoreUnannotated,
       'include_if_null': instance.includeIfNull,
       'nullable': instance.nullable,
     };