+++ json_serializable/test/shared_config.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:json_annotation/json_annotation.dart';
-import 'package:test/test.dart';
 
 final jsonSerializableFields = generatorConfigDefaultJson.keys.toList();
 
@@ -19,24 +18,7 @@ final generatorConfigNonDefaultJson =
   disallowUnrecognizedKeys: true,
   explicitToJson: true,
   fieldRename: FieldRename.kebab,
+  ignoreUnannotated: true,
   includeIfNull: false,
   nullable: false,
 ).toJson());
-
-// TODO(kevmoo): remove all of this logic once json_annotation v3 exists
-void expectMapMatchExcludingDeprecated(
-    Map<String, dynamic> actual, Map<String, dynamic> expected) {
-  Map<String, dynamic> exclude(Map<String, dynamic> source) => Map.fromEntries(
-      source.entries.where((e) => !deprecatedKeys.contains(e.key)));
-
-  expect(
-    exclude(actual),
-    exclude(expected),
-  );
-}
-
-const deprecatedKeys = {
-  'encode_empty_collection',
-  'generate_to_json_function',
-  'use_wrappers',
-};